requirements.txt
numpy==1.25.1
pandas==2.1.0
xlrd == 2.0.1
XlsxWriter == 3.1.3
python-dotenv == 1.0.0
pytz~=2023.3.post1
tinkoff-investments